backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for Todo application"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "sqlmodel>=0.0.19",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
]

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
dev-dependencies = ["pytest>=7.4.0", "pytest-asyncio>=0.23.0"]
